move classes from Diffview to multiple files to make the DiffView file smaller
So it is more comfortable to search anything
