+++ src/ui/DiffView/Line.h
@@ -1,0 +1,32 @@
+#ifndef LINE_H
+#define LINE_H
+
+class Line
+{
+public:
+  Line(char origin, int oldLine, int newLine)
+    : mOrigin(origin)
+  {
+    mOldLine = (oldLine >= 0) ? QByteArray::number(oldLine) : QByteArray();
+    mNewLine = (newLine >= 0) ? QByteArray::number(newLine) : QByteArray();
+  }
+
+  char origin() const { return mOrigin; }
+  QByteArray oldLine() const { return mOldLine; }
+  QByteArray newLine() const { return mNewLine; }
+
+  bool newline() const { return mNewline; }
+  void setNewline(bool newline) { mNewline = newline; }
+
+  int matchingLine() const { return mMatchingLine; }
+  void setMatchingLine(int line) { mMatchingLine = line; }
+
+private:
+  char mOrigin = -1;
+  bool mNewline = true;
+  int mMatchingLine = -1;
+  QByteArray mOldLine;
+  QByteArray mNewLine;
+};
+
+#endif // LINE_H
